+++ src/Provider/RequestLocaleProvider.php
@@ -15,8 +15,9 @@ namespace Sonata\TranslationBundle\Provider;
 
 use Sonata\TranslationBundle\Admin\Extension\AbstractTranslatableAdminExtension;
 use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Contracts\Service\ResetInterface;
 
-final class RequestLocaleProvider implements LocaleProviderInterface
+final class RequestLocaleProvider implements LocaleProviderInterface, ResetInterface
 {
     private ?string $translatableLocale = null;
 
@@ -33,6 +34,11 @@ final class RequestLocaleProvider implements LocaleProviderInterface
         }
 
         return $this->translatableLocale;
+    }
+
+    public function reset(): void
+    {
+        $this->translatableLocale = null;
     }
 
     private function getFromRequestOrDefault(): string