+++ src/Resources/config/service_gedmo.php
@@ -18,8 +18,10 @@ use Sonata\TranslationBundle\Admin\Extension\Gedmo\TranslatableAdminExtension;
 return static function (ContainerConfigurator $containerConfigurator): void {
     $containerConfigurator->services()
 
+        // @phpstan-ignore-next-line classConstant.internalClass
         ->set('sonata_translation.admin.extension.gedmo_translatable', TranslatableAdminExtension::class)
             ->tag('sonata.admin.extension')
+            ->tag('kernel.reset', ['method' => 'reset'])
             ->args([
                 service('sonata_translation.checker.translatable'),
                 service('sonata_translation.listener.translatable'),