+++ src/Resources/config/service_knplabs.php
@@ -20,8 +20,10 @@ use Sonata\TranslationBundle\Provider\Knplabs\LocaleProvider;
 return static function (ContainerConfigurator $containerConfigurator): void {
     $containerConfigurator->services()
 
+        // @phpstan-ignore-next-line classConstant.internalClass
         ->set('sonata_translation.admin.extension.knplabs_translatable', TranslatableAdminExtension::class)
             ->tag('sonata.admin.extension')
+            ->tag('kernel.reset', ['method' => 'reset'])
             ->args([
                 service('sonata_translation.checker.translatable'),
                 service('sonata_translation.admin.provider.request_locale_provider'),