+++ tests/Admin/Extension/AbstractTranslatableAdminExtensionTest.php
@@ -18,6 +18,9 @@ use Sonata\AdminBundle\Admin\AdminInterface;
 use Sonata\TranslationBundle\Admin\Extension\AbstractTranslatableAdminExtension;
 use Sonata\TranslationBundle\Checker\TranslatableChecker;
 use Sonata\TranslationBundle\Provider\LocaleProviderInterface;
+use Sonata\TranslationBundle\Provider\RequestLocaleProvider;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RequestStack;
 
 final class AbstractTranslatableAdminExtensionTest extends TestCase
 {
@@ -47,5 +50,34 @@ final class AbstractTranslatableAdminExtensionTest extends TestCase
         $parameters = $this->extension->configurePersistentParameters(static::createStub(AdminInterface::class), []);
 
         static::assertSame(['tl' => 'es'], $parameters);
+    }
+
+    public function testReset(): void
+    {
+        $request = new Request();
+        $request->query->set(AbstractTranslatableAdminExtension::TRANSLATABLE_LOCALE_PARAMETER, 'cs');
+        $requestStack = new RequestStack();
+        $requestStack->push($request);
+        $localeProvider = new RequestLocaleProvider($requestStack, 'en');
+
+        $translatableChecker = new TranslatableChecker();
+
+        $extension = new
+            /**
+             * @template-extends AbstractTranslatableAdminExtension<object>
+             */
+            class($translatableChecker, $localeProvider) extends AbstractTranslatableAdminExtension {
+                public function getLocalePublicly(): string
+                {
+                    return $this->getTranslatableLocale();
+                }
+            };
+
+        static::assertSame('cs', $extension->getLocalePublicly());
+        $request->query->set(AbstractTranslatableAdminExtension::TRANSLATABLE_LOCALE_PARAMETER, 'es');
+        $localeProvider->reset();
+        static::assertSame('cs', $extension->getLocalePublicly());
+        $extension->reset();
+        static::assertSame('es', $extension->getLocalePublicly());
     }
 }